+++ src/repository.ts
@@ -2,13 +2,7 @@
 /**
  * リポジトリ
  */
-import { MongoRepository as PerformanceRepo } from './repo/performance';
 import { MongoRepository as ReportRepo } from './repo/report';
-
-/**
- * パフォーマンスリポジトリ
- */
-export class Performance extends PerformanceRepo { }
 
 /**
  * レポートリポジトリ