+++ CHANGELOG.md
@@ -10,9 +10,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- @alverca/factoryで再構築
+
 ### Deprecated
 
 ### Removed
+
+- パフォーマンスリポジトリを削除
 
 ### Fixed